requirements.txt
streamlit
langchain>=0.1.0
langchain-community>=0.0.10
langgraph
openai
faiss-cpu  # GPU 사용시 faiss-gpu
pdfplumber
python-dotenv
tiktoken
numpy
watchdog
flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
Werkzeug==3.0.1